Fix nvidia-driver-515.postinst install on BIOS
Bios machines dont have a /boot/efi, thus df | grep /boot/efi
fails, which causes the script to fail, which causes the nvidia
drivers not to install.

This change just checks for the return value of grep, and if there
is an error we return 0 so that the postinstall script exits
gracefully.

Thanks @mikejaques for finding this bug and reporting it so quickly
on our discord!